Updates on mediaContainer
