+++ app/components/Questionnaire/index.js
@@ -2,7 +2,7 @@ import QuestionAnswers from './QuestionAnswers';
 import QuestionOverview from './QuestionOverview';
 import Answers from './Answers';
 import Question from './Question';
-import ImageContainer from './ImageContainer';
+import MediaContainer from './MediaContainer';
 import PrefilledAnswerText from './PrefilledAnswerText';
 
-export { QuestionAnswers, QuestionOverview, Answers, Question, ImageContainer, PrefilledAnswerText };
+export { QuestionAnswers, QuestionOverview, Answers, Question, MediaContainer, PrefilledAnswerText };
